Update to 2021 WPILib
Followed the manual import instructions
(<https://docs.wpilib.org/en/stable/docs/software/vscode-overview/importing-gradle-project.html#manual-import-process-2020-to-2021>)
for importing a 2020 project to 2021 WPILib. The codebase will need
imported to the 2022 version of WPILib after it's out of beta, I'm
assuming that's during the start of build season and so will be done
with the creation of the official season repository.